refactor: move schema into data_layer
