+++ src/data_layer/public/KnexMigrationsLock.ts
@@ -2,7 +2,9 @@
 // This file is automatically generated by Kanel. Do not modify manually.
 
 /** Identifier type for public.knex_migrations_lock */
-export type KnexMigrationsLockIndex = number & { __brand: 'KnexMigrationsLockIndex' };
+export type KnexMigrationsLockIndex = number & {
+  __brand: 'KnexMigrationsLockIndex';
+};
 
 /** Represents the table public.knex_migrations_lock */
 export default interface KnexMigrationsLock {
